telegram-bot/requirements.txt
websockets==9.1
requests==2.26.0
python-telegram-bot==13.11
getmac==0.8.3
bs4==0.0.1
builtwith==1.3.4
beautifulsoup4==4.10.0
sockets==1.0.0
